# docs/api/ai-api.openapi.json
{
  "components": {
    "schemas": {
...
      "AdminUsageResponse": {
        "description": "Every account's counter of one UTC day, most tokens first.",
        "properties": {
          "day": {
            "title": "Day",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminUsageItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "day",
          "items"
        ],
        "title": "AdminUsageResponse",
        "type": "object"
      },
...
      "CityStatsResponse": {
        "properties": {
          "city": {
            "title": "City",
            "type": "string"
          },
          "errors": {
            "title": "Errors",
            "type": "integer"
          },
          "turns": {
            "title": "Turns",
            "type": "integer"
          }
        },
        "required": [
          "city",
          "turns",
          "errors"
        ],
        "title": "CityStatsResponse",
        "type": "object"
      },
...
      "DayStatsResponse": {
        "properties": {
          "cancelled": {
            "title": "Cancelled",
            "type": "integer"
          },
          "cost_usd": {
            "title": "Cost Usd",
            "type": "number"
          },
          "day": {
            "title": "Day",
            "type": "string"
          },
          "embed_tokens": {
            "title": "Embed Tokens",
            "type": "integer"
          },
          "errors": {
            "title": "Errors",
            "type": "integer"
          },
          "first_event_p50_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Event P50 Ms"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "latency_p50_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency P50 Ms"
          },
          "latency_p95_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency P95 Ms"
          },
          "ok": {
            "title": "Ok",
            "type": "integer"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "turns": {
            "title": "Turns",
            "type": "integer"
          }
        },
        "required": [
          "day",
          "turns",
          "ok",
          "errors",
          "cancelled",
          "input_tokens",
          "output_tokens",
          "embed_tokens",
          "cost_usd",
          "latency_p50_ms",
          "latency_p95_ms",
          "first_event_p50_ms"
        ],
        "title": "DayStatsResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KindStatsResponse": {
        "properties": {
          "cost_usd": {
            "title": "Cost Usd",
            "type": "number"
          },
          "errors": {
            "title": "Errors",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "turns": {
            "title": "Turns",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "turns",
          "errors",
          "cost_usd"
        ],
        "title": "KindStatsResponse",
        "type": "object"
      },
      "ModelStatsResponse": {
        "properties": {
          "cost_usd": {
            "title": "Cost Usd",
            "type": "number"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "turns": {
            "title": "Turns",
            "type": "integer"
          }
        },
        "required": [
          "model",
          "turns",
          "input_tokens",
          "output_tokens",
          "cost_usd"
        ],
        "title": "ModelStatsResponse",
        "type": "object"
      },
      "NeverUsedDocResponse": {
        "properties": {
          "doc_id": {
            "title": "Doc Id",
            "type": "string"
          },
          "retrieved": {
            "title": "Retrieved",
            "type": "integer"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "doc_id",
          "title",
          "retrieved"
        ],
        "title": "NeverUsedDocResponse",
        "type": "object"
      },
...
      "TraceStatsResponse": {
        "description": "The stats of a range of days; definitions in `application/trace_stats.py`.",
        "properties": {
          "by_city": {
            "items": {
              "$ref": "#/components/schemas/CityStatsResponse"
            },
            "title": "By City",
            "type": "array"
          },
          "by_kind": {
            "items": {
              "$ref": "#/components/schemas/KindStatsResponse"
            },
            "title": "By Kind",
            "type": "array"
          },
          "by_model": {
            "items": {
              "$ref": "#/components/schemas/ModelStatsResponse"
            },
            "title": "By Model",
            "type": "array"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/DayStatsResponse"
            },
            "title": "Days",
            "type": "array"
          },
          "end": {
            "title": "End",
            "type": "string"
          },
          "never_used": {
            "items": {
              "$ref": "#/components/schemas/NeverUsedDocResponse"
            },
            "title": "Never Used",
            "type": "array"
          },
          "rag": {
            "$ref": "#/components/schemas/RagStatsResponse"
          },
          "start": {
            "title": "Start",
            "type": "string"
          },
          "top_used": {
            "items": {
              "$ref": "#/components/schemas/UsedDocResponse"
            },
            "title": "Top Used",
            "type": "array"
          },
          "totals": {
            "$ref": "#/components/schemas/TotalsResponse"
          }
        },
        "required": [
          "start",
          "end",
          "days",
          "totals",
          "by_kind",
          "by_model",
          "by_city",
          "rag",
          "top_used",
          "never_used"
        ],
        "title": "TraceStatsResponse",
        "type": "object"
      },
...
      "TurnDetailResponse": {
        "description": "One turn, whole: summary, context, steps by `seq`, SSE timeline.",
        "properties": {
          "context": {
            "$ref": "#/components/schemas/TurnContextResponse"
          },
          "spans": {
            "items": {
              "$ref": "#/components/schemas/SpanResponse"
            },
            "title": "Spans",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/TurnSummaryResponse"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/EventMarkResponse"
            },
            "title": "Timeline",
            "type": "array"
          }
        },
        "required": [
          "summary",
          "context",
          "spans",
          "timeline"
        ],
        "title": "TurnDetailResponse",
        "type": "object"
      },
      "TurnPageResponse": {
        "description": "A page of turns; pass `next_cursor` back to read the next one.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TurnSummaryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "TurnPageResponse",
        "type": "object"
      },
      "TurnSummaryResponse": {
        "description": "One turn as a list shows it; `day` and `sk` are its summary's key.",
        "properties": {
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "answer_preview": {
            "title": "Answer Preview",
            "type": "string"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "day": {
            "title": "Day",
            "type": "string"
          },
          "docs_retrieved": {
            "title": "Docs Retrieved",
            "type": "integer"
          },
          "docs_used": {
            "title": "Docs Used",
            "type": "integer"
          },
          "dropped_ids": {
            "title": "Dropped Ids",
            "type": "integer"
          },
          "embed_tokens": {
            "title": "Embed Tokens",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "events": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Events",
            "type": "object"
          },
          "first_event_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Event Ms"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "kind": {
            "enum": [
              "planner",
              "chat",
              "card"
            ],
            "title": "Kind",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "llm_calls": {
            "title": "Llm Calls",
            "type": "integer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "ops": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Ops",
            "type": "object"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "prices_stripped": {
            "title": "Prices Stripped",
            "type": "integer"
          },
          "pricing_version": {
            "title": "Pricing Version",
            "type": "string"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "question_preview": {
            "title": "Question Preview",
            "type": "string"
          },
          "repairs": {
            "title": "Repairs",
            "type": "integer"
          },
          "retrievals": {
            "title": "Retrievals",
            "type": "integer"
          },
          "route": {
            "title": "Route",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "sk": {
            "title": "Sk",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/RetrievedDocResponse"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "enum": [
              "ok",
              "error",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "trip_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Id"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          },
          "turn_id": {
            "title": "Turn Id",
            "type": "string"
          },
          "warnings": {
            "title": "Warnings",
            "type": "integer"
          }
        },
        "required": [
          "turn_id",
          "ts",
          "day",
          "sk",
          "kind",
          "route",
          "subject",
          "session_id",
          "trip_id",
          "city",
          "language",
          "action",
          "model",
          "provider",
          "prompt_version",
          "input_tokens",
          "output_tokens",
          "embed_tokens",
          "cost_usd",
          "pricing_version",
          "latency_ms",
          "first_event_ms",
          "status",
          "error_code",
          "llm_calls",
          "retrievals",
          "docs_retrieved",
          "docs_used",
          "repairs",
          "dropped_ids",
          "prices_stripped",
          "warnings",
          "events",
          "ops",
          "sources",
          "question_preview",
          "answer_preview",
          "truncated"
        ],
        "title": "TurnSummaryResponse",
        "type": "object"
      },
...
      "UsedDocResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "doc_id": {
            "title": "Doc Id",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "doc_id",
          "title",
          "count"
        ],
        "title": "UsedDocResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Kyrian World \u2014 AI API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/ai/admin/sessions/{session_id}": {
      "get": {
        "description": "One planner conversation's turns, oldest first.",
        "operationId": "read_session_api_v1_ai_admin_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Session",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/ai/admin/stats": {
      "get": {
        "description": "Per-day and total counts, tokens, cost, latency, and the RAG metrics\nover `[start, end]` (at most 31 days).",
        "operationId": "read_stats_api_v1_ai_admin_stats_get",
        "parameters": [
          {
            "description": "First day, `YYYY-MM-DD`.",
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "description": "First day, `YYYY-MM-DD`.",
              "format": "date",
              "title": "Start",
              "type": "string"
            }
          },
          {
            "description": "Last day, included.",
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "description": "Last day, included.",
              "format": "date",
              "title": "End",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceStatsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Stats",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/ai/admin/turns": {
      "get": {
        "description": "Turns by session (oldest first), else by day, else by user (newest\nfirst), with the other filters applied on top.",
        "operationId": "list_turns_api_v1_ai_admin_turns_get",
        "parameters": [
          {
            "description": "A UTC day, `YYYY-MM-DD`.",
            "in": "query",
            "name": "day",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A UTC day, `YYYY-MM-DD`.",
              "title": "Day"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "planner",
                    "chat",
                    "card"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ok",
                    "error",
                    "cancelled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "description": "The token subject of a user.",
            "in": "query",
            "name": "subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The token subject of a user.",
              "title": "Subject"
            }
          },
          {
            "description": "A planner session id.",
            "in": "query",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A planner session id.",
              "title": "Session"
            }
          },
          {
            "in": "query",
            "name": "trip_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trip Id"
            }
          },
          {
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Turns",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/ai/admin/turns/{turn_id}": {
      "get": {
        "description": "One turn, whole: summary, request context, steps and SSE timeline.",
        "operationId": "read_turn_api_v1_ai_admin_turns__turn_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "turn_id",
            "required": true,
            "schema": {
              "title": "Turn Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Turn",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/ai/admin/usage": {
      "get": {
        "description": "Every account's token counter of one UTC day (today by default), most\ntokens first. `tokens` is input + output, what the daily limit counts.",
        "operationId": "read_usage_api_v1_ai_admin_usage_get",
        "parameters": [
          {
            "description": "A UTC day, `YYYY-MM-DD`; today.",
            "in": "query",
            "name": "day",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A UTC day, `YYYY-MM-DD`; today.",
              "title": "Day"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUsageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Usage",
        "tags": [
          "Admin"
        ]
      }
    },
...
  }
}